Adding domain validations and exceptions

+++ src/main/java/com/glaydson/devicesapi/service/DeviceService.java
@@ -1,8 +1,11 @@
+// src/main/java/com/glaydson/devicesapi/service/DeviceService.java
 package com.glaydson.devicesapi.service;
 
+import com.glaydson.devicesapi.exception.DeviceInUseException;
+import com.glaydson.devicesapi.exception.InvalidDeviceStateException;
+import com.glaydson.devicesapi.exception.ResourceNotFoundException;
 import com.glaydson.devicesapi.model.Device;
 import com.glaydson.devicesapi.repository.DeviceRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,14 +14,16 @@ import java.util.Optional;
 @Service
 public class DeviceService {
 
-    private DeviceRepository deviceRepository;
+    private final DeviceRepository deviceRepository;
 
-    // Contructor
     public DeviceService(DeviceRepository deviceRepository) {
         this.deviceRepository = deviceRepository;
     }
 
     public Device createDevice(Device device) {
+        if (device.getState() == null) {
+            throw new InvalidDeviceStateException("Device state cannot be null");
+        }
         return deviceRepository.save(device);
     }
 
@@ -39,10 +44,30 @@ public class DeviceService {
     }
 
     public Device updateDevice(Device device) {
+        Device existingDevice = deviceRepository.findById(device.getId())
+                .orElseThrow(() -> new ResourceNotFoundException("Device not found for this id :: " + device.getId()));
+
+        if (existingDevice.getState() == Device.State.IN_USE) {
+            if (!existingDevice.getName().equals(device.getName()) || !existingDevice.getBrand().equals(device.getBrand())) {
+                throw new DeviceInUseException("Name and brand cannot be updated if the device is in use");
+            }
+        }
+
+        if (device.getCreationTime() != null && !device.getCreationTime().equals(existingDevice.getCreationTime())) {
+            throw new InvalidDeviceStateException("Creation time cannot be updated");
+        }
+
         return deviceRepository.save(device);
     }
 
     public void deleteDevice(Long id) {
+        Device existingDevice = deviceRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Device not found for this id :: " + id));
+
+        if (existingDevice.getState() == Device.State.IN_USE) {
+            throw new DeviceInUseException("Devices in use cannot be removed");
+        }
+
         deviceRepository.deleteById(id);
     }
 }
